Moved add contact call in userdetailpresenter to usersinteractor
Cleaned up both interactor and presenter levels while I was at it
