+++ app/src/main/java/com/bookyrself/bookyrself/presenters/UserDetailPresenter.java
@@ -1,5 +1,6 @@
 package com.bookyrself.bookyrself.presenters;
 
+import com.bookyrself.bookyrself.interactors.ContactsInteractor;
 import com.bookyrself.bookyrself.interactors.EventsInteractor;
 import com.bookyrself.bookyrself.interactors.UsersInteractor;
 import com.bookyrself.bookyrself.models.SerializedModels.EventDetail.EventDetail;
@@ -17,9 +18,8 @@ import retrofit2.Response;
  * Created by benmedcalf on 1/13/18.
  */
 
-public class UserDetailPresenter implements EventsInteractor.EventsInteractorListener, UsersInteractor.UserDetailInteractorListener {
+public class UserDetailPresenter implements EventsInteractor.EventsInteractorListener, UsersInteractor.UserDetailInteractorListener{
     private final UserDetailPresenterListener listener;
-    private final FirebaseService service;
     private final EventsInteractor eventsInteractor;
     private final UsersInteractor usersInteractor;
 
@@ -29,7 +29,6 @@ public class UserDetailPresenter implements EventsInteractor.EventsInteractorLis
 
     public UserDetailPresenter(UserDetailPresenterListener listener) {
         this.listener = listener;
-        this.service = new FirebaseService();
         this.eventsInteractor = new EventsInteractor(this);
         this.usersInteractor = new UsersInteractor(this);
     }
@@ -41,22 +40,8 @@ public class UserDetailPresenter implements EventsInteractor.EventsInteractorLis
         usersInteractor.getUserDetails(userId);
     }
 
-
-    //TODO: Put this in the interactor
-    public void addContactToUser(String userId, String contactId) {
-        HashMap<String, Boolean> request = new HashMap<>();
-        request.put(contactId, true);
-        service.getAPI().addContactToUser(request, userId).enqueue(new Callback<HashMap<String, Boolean>>() {
-            @Override
-            public void onResponse(Call<HashMap<String, Boolean>> call, Response<HashMap<String, Boolean>> response) {
-                listener.presentSuccess("added to contacts!");
-            }
-
-            @Override
-            public void onFailure(Call<HashMap<String, Boolean>> call, Throwable t) {
-                presentError(t.getMessage());
-            }
-        });
+    public void addContactToUser(String contactId, String userId) {
+        usersInteractor.addContactToUser(contactId,userId);
     }
 
     @Override
@@ -79,7 +64,7 @@ public class UserDetailPresenter implements EventsInteractor.EventsInteractorLis
 
     @Override
     public void contactSuccessfullyAdded() {
-
+        listener.presentSuccess("Added contact!");
     }
