+++ app/src/main/res/drawable/ic_bio.xml
@@ -1,0 +1,4 @@
+<vector android:height="24dp" android:viewportHeight="511.998"
+    android:viewportWidth="511.998" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="M503.169,147.785L461,128.758V99.797c0,-5.889 -3.446,-11.233 -8.81,-13.663c-30.339,-13.745 -65.33,-21.01 -101.19,-21.01c-33.446,0 -66.135,6.319 -95,18.322c-28.865,-12.003 -61.554,-18.322 -95,-18.322c-35.86,0 -70.851,7.265 -101.19,21.01C54.446,88.564 51,93.909 51,99.797v28.961L8.831,147.785C3.456,150.21 0,155.56 0,161.457v260.399c0,8.53 4.278,16.385 11.444,21.011c7.062,4.56 15.971,5.333 23.835,1.782l56.818,-25.639c20.15,-9.09 43.839,-13.894 68.503,-13.894c49.858,0 76.132,18.527 89.23,23.242c3.923,1.77 8.417,1.77 12.339,0c13.035,-4.692 39.404,-23.242 89.23,-23.242c24.664,0 48.352,4.804 68.5,13.893l56.815,25.637c7.767,3.507 16.68,2.844 23.84,-1.779c7.166,-4.626 11.444,-12.48 11.444,-21.011V161.457C512,155.56 508.544,150.21 503.169,147.785zM241,337.85c-24.976,-8.353 -52.223,-12.726 -80,-12.726c-27.775,0 -55.025,4.357 -80,12.709c0,-6.938 0,-216.153 0,-228.093c49.594,-19.479 110.407,-19.469 160,0.012V337.85zM431,337.834c-24.975,-8.352 -52.225,-12.709 -80,-12.709c-27.777,0 -55.024,4.373 -80,12.726v-228.1c49.593,-19.481 110.406,-19.491 160,-0.011C431,117.109 431,317.095 431,337.834z"/>
+</vector>
